+++ next.config.ts
@@ -9,6 +9,7 @@ const nextConfig: NextConfig = {
       },
     ],
   },
+  poweredByHeader: false,
 };
 
 export default nextConfig;